Speed up xtrabackup prepare phase
As announced (https://www.percona.com/blog/making-your-mysql-backup-process-faster-introducing-percona-xtrabackup-smart-memory-estimation/),
xtrabackup has a new feature which significantly speeds up basebackup
preparation operations.

This is disabled by default, but will soon be enabled.

Because the progress metric is not updated during this phase, this
change will reduce the number of false-positive escalations related to
restore operations not progressing.