Mark streams opened for old master's backups as closed [BF-2225]
When the role of the current node is master, then we mark any observe
streams as closed, as those are assumed to be opened for backups
initiated by the old master after the promotion process of the current
node was triggered. This is necessary so that new backups can be
requested, as we do not allow creating new streams if 2 or more are
already opened.

[BF-2225]